requirements.txt

fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
psycopg2==2.9.2
SQLAlchemy==1.4.25
pytest==6.2.5
alembic
pymongo
requests
python-dotenv
schedule
pandas
urllib3
colorlog
apache-airflow-providers-docker
pytest-mock